+++ test/case-rmnoname.sh
@@ -1,3 +1,3 @@
 
-./xsvutils ./test/data/rmnoname.tsv rmnoname -v4 > ./var/test/rmnoname-1.tsv
+./xsvutils ./test/data/rmnoname.tsv rmnoname > ./var/test/rmnoname-1.tsv