+++ CHANGELOG.md
@@ -8,6 +8,7 @@ To see compatibility information, `xsvutils help compatibility`.
 - add support input of xz file
 - `-` means stdin
 - add input file name validation
+- remove -v4 option of `rmnoname` subcommand
 
 ## version 0.4 (2018/07/10)
 - support -v1,-v2,-v3,-v4